Refactor: Extract shared locale parsing utility and reorganize MockPurchasesType
- Create shared createLocaleFromString() function in LocaleHelpers.kt
  - Supports both "es-ES" and "es_ES" locale formats
  - Eliminates duplicate code across PaywallViewModel and CustomerCenterViewModel
- Move MockPurchasesType from main to test/debug source sets
  - Follows project patterns for mock organization
  - Available in test and debug builds, not production
- Update PaywallState to use shared locale parsing utility
- Fix import paths for proper dependency resolution
