+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/data/PurchasesType.kt
@@ -102,45 +102,3 @@ internal class PurchasesImpl(private val purchases: Purchases = Purchases.shared
     override val preferredUILocaleOverride: String?
         get() = purchases.preferredUILocaleOverride
 }
-
-/**
- * Mock implementation of [PurchasesType] for tests and previews
- */
-internal class MockPurchasesType(
-    override val preferredUILocaleOverride: String? = null,
-    override val purchasesAreCompletedBy: PurchasesAreCompletedBy = PurchasesAreCompletedBy.REVENUECAT,
-    override val storefrontCountryCode: String? = null,
-    override val customerCenterListener: CustomerCenterListener? = null,
-) : PurchasesType {
-    override suspend fun awaitPurchase(purchaseParams: PurchaseParams.Builder): PurchaseResult {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override suspend fun awaitRestore(): CustomerInfo {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override suspend fun awaitOfferings(): Offerings {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override suspend fun awaitCustomerInfo(fetchPolicy: CacheFetchPolicy): CustomerInfo {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override suspend fun awaitCustomerCenterConfigData(): CustomerCenterConfigData {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override suspend fun awaitGetProduct(productId: String, basePlan: String?): StoreProduct? {
-        throw NotImplementedError("Mock implementation")
-    }
-
-    override fun track(event: FeatureEvent) {
-        // No-op for mock
-    }
-
-    override fun syncPurchases() {
-        // No-op for mock
-    }
-}
